re-struct env
