+++ minesrl/envs/__init__.py
@@ -1,1 +1,2 @@
 from .mines_env import MinesEnv
+from .mines_console_env import MinesConsoleEnv
