+++ minesrl/game/__init__.py
@@ -1,2 +1,2 @@
-from .mines_core import MinesGame
+from .mines_game import MinesGame
 from .mines_console import MinesConsole
