+++ minesrl/game/mines_console.py
@@ -1,31 +1,11 @@
 from minesrl.game import *
+from minesrl.envs import MinesConsoleEnv
 
 
 class MinesConsole(MinesGame):
     def __init__(self, row=16, col=16, mine=40):
-        super().__init__(row, col, mine)
-
-    def render(self):
-        map = self.env.observation_space
-        print('   |', end='')
-        for c in range(self.env.cols):
-            print('{0:2d}'.format(c % 10), end='')
-        print(' ')
-        for c in range(self.env.cols + 2):
-            print('--', end='')
-        print(' ')
-        for r in range(self.env.rows):
-            print('{0:3d}| '.format(r), end='')
-            for c in range(self.env.cols):
-                if map[r][c] == self.env.UNKNOWN:
-                    print('.', end=' ')
-                elif map[r][c] == self.env.MINE:
-                    print('💣', end=' ')
-                elif map[r][c] == 0:
-                    print(' ', end=' ')
-                else:
-                    print(map[r][c], end=' ')
-            print()
+        env = MinesConsoleEnv(row, col, mine)
+        super().__init__(env)
 
     def get_input(self):
         inp = input('Action: ')
